PointProcessFilters

Documentation for PointProcessFilters. A package for working filtering point processes. Currently only provides bandpass filters for a variety of spatial regions.

Base.filter

PointProcessFilters.boxpass

PointProcessFilters.centeredboxpass

PointProcessFilters.centeredcirclepass

PointProcessFilters.check_region_condition

PointProcessFilters.circlepass

PointProcessFilters.impulse_response

PointProcessFilters.impulseresponse_precomp

PointProcessFilters.ringpass

Base.filter — Method

filter(X::PointSet{D,T}, r::SpatialFilter{D}) where {D,T}

Applies a given spatial filter r to a point set. The output can be evaluated as a function.

Example

r = 1.0
X = PointSet(Point.(rand(10), rand(10)))
Y = filter(X, r)
Y(1.0,1.0)
u = (1.0,1.0)
Y(u)

PointProcessFilters.boxpass — Method

boxpass(X::PointSet, a, b)

Applies the box pass filter to boxes with corners a, b and -a, -b.

PointProcessFilters.centeredboxpass — Method

centeredboxpass(X::PointSet, a)

Applies a centered box pass filter, i.e. a box with corners -a, a.

PointProcessFilters.centeredcirclepass — Method

centeredcirclepass(X::PointSet{D,T}, r) where {D,T}

Applies a circle pass filter centered at zero, with radius r.

PointProcessFilters.check_region_condition — Method

check_region_condition(::Geometry)

Perform additional checks on the provided region.

PointProcessFilters.circlepass — Method

circlepass(X::PointSet, p, r)

Applies a circle pass filter with circles centered at p and -p in wavenumber and radii of r.

PointProcessFilters.impulse_response — Method

impulse_response(r::RegionPassFilter)

Computes the impulse response function for a region pass filter. Result is of type ImpulseResponseRegion, a callable struct.

PointProcessFilters.impulseresponse_precomp — Method

impulseresponse_precomp(::Geometry)

Generates precomputed quantities for a given region.

PointProcessFilters.ringpass — Method

ringpass(X::PointSet{D,T}, a, b) where {D,T}

Applies a ring pass (annulus) filter with the inner radius a and outer radius b.