Available Regions

The bandpass filtering will filter frequencies within a given region. There are a variety of provided regions, in particular:

Ball

PointProcessFilters.circlepass — Function

circlepass(X::PointSet, p, r)

Applies a circle pass filter with circles centered at p and -p in wavenumber and radii of r.

Centered Ball

PointProcessFilters.centeredcirclepass — Function

centeredcirclepass(X::PointSet{D,T}, r) where {D,T}

Applies a circle pass filter centered at zero, with radius r.

Ring

PointProcessFilters.ringpass — Function

ringpass(X::PointSet{D,T}, a, b) where {D,T}

Applies a ring pass (annulus) filter with the inner radius a and outer radius b.

Box

PointProcessFilters.boxpass — Function

boxpass(X::PointSet, a, b)

Applies the box pass filter to boxes with corners a, b and -a, -b.

Centered Box

PointProcessFilters.centeredboxpass — Function

centeredboxpass(X::PointSet, a)

Applies a centered box pass filter, i.e. a box with corners -a, a.